feat: restructure application layout and enhance feature organization
- Update layout.tsx to use absolute imports for components and improve structure
- Remove deprecated page.tsx and integrate marketing sections into the new layout
- Create dashboard layout and pages with responsive design for user metrics
- Implement overlay configuration system with real-time preview and settings
- Introduce new components for authentication and dashboard features
- Optimize imports and file organization for better maintainability
