+++ app/(dashboard)/layout.tsx
@@ -1,0 +1,11 @@
+export default function DashboardLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return (
+    <div className="min-h-screen bg-pearl">
+      {children}
+    </div>
+  )
+}
