+++ app/(marketing)/layout.tsx
@@ -1,0 +1,14 @@
+import { ConditionalNavbar } from '@/components/layout/ConditionalNavbar'
+
+export default function MarketingLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return (
+    <>
+      <ConditionalNavbar />
+      {children}
+    </>
+  )
+}
