+++ components/ui/KickIcon.tsx
