+++ features/auth/index.ts
@@ -1,0 +1,15 @@
+// Auth feature exports
+export { LoginForm } from './components/LoginForm'
+export { RegisterForm } from './components/RegisterForm'
+export { AuthErrorBoundary } from './components/AuthErrorBoundary'
+
+// Hooks
+export { useAuth } from './hooks/useAuth'
+export { useAuthState } from './hooks/useAuthQueries'
+
+// Services
+export * from './services/auth'
+export * from './services/secure-auth'
+
+// Types
+export * from './types/auth'
