+++ features/dashboard/components/WeeklyStatsCard.tsx
