+++ features/dashboard/index.ts
@@ -1,0 +1,9 @@
+// Dashboard feature exports
+export { DashboardHeader } from './components/DashboardHeader'
+export { TopDonorsCard } from './components/TopDonorsCard'
+export { LiveStatusCard } from './components/LiveStatusCard'
+export { QuickActionsCard } from './components/QuickActionsCard'
+export { QuickSetupSection } from './components/QuickSetupSection'
+export { WeeklyStatsCard } from './components/WeeklyStatsCard'
+export { PlatformSection } from './components/PlatformSection'
+export { RecentDonationsFeed } from './components/RecentDonationsFeed'
