+++ features/marketing/index.ts
@@ -1,0 +1,4 @@
+// Marketing feature exports
+export { HeroSection } from './components/HeroSection'
+export { HowItWorksSection } from './components/HowItWorksSection'
+export { FinalCTASection } from './components/FinalCTASection'
