+++ features/overlay/index.ts
@@ -1,0 +1,4 @@
+// Overlay feature exports
+export { OverlayLinkCard } from './components/OverlayLinkCard'
+export { OverlayPreview } from './components/OverlayPreview'
+export { OverlaySettings } from './components/OverlaySettings'
